rossanodan-issue-35: removed tailwind css and installed bootstrap; re-built coming soon component and changed naming convention to be more consistent and clear
