+++ package.json
@@ -6,15 +6,16 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "bootstrap": "^4.5.2",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
+    "react-bootstrap": "^1.3.0",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3"
   },
   "scripts": {
-    "start": "npm run build:style && react-scripts start",
-    "build:style": "tailwind build src/styles/index.css -o src/styles/tailwind.css",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "jest",
     "lint": "./node_modules/.bin/eslint --ignore-path .gitignore ."
@@ -40,11 +41,8 @@
     ]
   },
   "devDependencies": {
-    "autoprefixer": "^10.0.0",
     "eslint-plugin-import": "^2.22.0",
     "eslint-plugin-react": "^7.20.6",
-    "husky": "^4.3.0",
-    "postcss-cli": "^7.1.2",
-    "tailwindcss": "^1.8.10"
+    "husky": "^4.3.0"
   }
 }
