--- postcss.config.js
@@ -1,8 +1,0 @@
-const tailwindcss = require('tailwindcss');
-// eslint-disable-next-line no-undef
-module.exports = {
-  plugins: [
-    tailwindcss('./tailwind.js'),
-    require('autoprefixer'),
-  ],
-};
