+++ src/coming-soon/coming-soon.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const ComingSoon = () => (
+  <>
+    <h1>Coming soon</h1>
+    <p>Don&apos;t worry, amazing things are on the way. Stay tuned!</p>
+  </>
+);
+
+export default ComingSoon;
