--- src/feature-comingsoon/comingsoon.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-const ComingSoon = () => (
-  <div className="h-screen w-screen bg-gray-100 flex items-center">
-    <div className="container flex flex-col md:flex-row items-center justify-center px-5 text-gray-700">
-        <div className="max-w-md">
-          <div className="text-5xl font-dark font-bold">Coming soon</div>
-            <br />
-            <p className="mb-8">But dont worry, awesome things are on the way.</p>
-      </div>
-    </div>
-  </div>
-);
-
-export default ComingSoon;
